Trip Service feature

+++ trip/src/main/java/com/erick/trip/domain/passenger/TripRequestPassenger.java
@@ -6,7 +6,7 @@ import lombok.ToString;
 @Data
 @ToString
 public class TripRequestPassenger {
-    private Long idPassenger;
+    private Integer idPassenger;
     private String addressOrigin;
     private String addressDestiny;
     private String district;
